fix: Article의 생성 시간 타입 변경
- 확장성을 고려하여 기존의 String 타입에서 LocalDateTime으로 변경하였다.

+++ src/main/java/codesquad/springcafe/exception/ArticleNotFoundException.java
@@ -1,7 +1,7 @@
 package codesquad.springcafe.exception;
 
 public class ArticleNotFoundException extends Exception {
-    public ArticleNotFoundException(int articleId) {
+    public ArticleNotFoundException(long articleId) {
         super(articleId + "라는 게시글은 존재하지 않습니다.");
     }
 }
